ai_things/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
future==0.18.3
idna==3.4
python-dotenv==1.0.0
requests==2.30.0
urllib3==1.26.6
